DUOS-1293 removed old member console, renamed NewMemberConsole to MemberConsole
